Initial landlock support.
Landlock is inferior to plegde/unveil. It has the usual issue of trying
to sandbox everything with the result that it only works if the sandbox is
as big as the world it runs in. It has no smarts about program execution, no
helpers to make basic things like DNS lookups simple, no smarts to help
creating portable rulesets that just work. Instead it includes a complicated
ABI with a fail-open by default design.

For rpki-client this means that process that are extensivly locked down
on OpenBSD get a lot of rights to work with landlock. e.g. the httpd process
gets full read access to all of your files. Same for the rsync process
which just needs to start rsync processes. Again full read/execute rights
need to be handed out to a process that is totally locked down to one file
in OpenBSD.

Maybe this is better than nothing but it is sad to see yet another failed
method of restricted-service mode that is hard to use properly. It requires
the developer or user to know all possible file access patterns of the tool
chain including all file accesses caused by libraries for all possible cases.
Sadly that is just impossible so one needs to open up way more than wanted.

Landlock probably needs FS_REFER to work so this is not good enough for
ABI 1 landlock.