Landlock ABI 1 is just broken and can not be enabled in the master process.
At the end of the run files are moved from their temporary location into
the validated cache but Landlock does not allow that because moving a file
out of a directory is considered a privilege escalation. So ABI 1 is just
unfit. In ABI 2 the LANDLOCK_ACCESS_FS_REFER flag was introduced to fix this.

The fix is rather ugly and just ignores unveil if "rwc" permissions are used.
In rpki-client only the master process needs these permissions and they are
also the first and only permisssion used in this case.