configure.ac: check for fts.h.
the configure script checks for fts functions and if found, adds the -lfts to
the generated makefiles but `fts.h' is directly included without any guards and
the rest of the code does use fts_* functions, rendering the test useless.

in systems like musl hosts where an external fts implementation needs to be
brought in, this causes the following error:

 repo.c:28:10: fatal error: 'fts.h' file not found
 #include <fts.h>

on musl hosts the error can be prevented by building and linking against
musl-fts (https://github.com/void-linux/musl-fts) but the configure script
should die if fts.h is not present during the configuration stage in order
to prevent surprises.